+++ leshan-core/src/main/java/org/eclipse/leshan/core/node/codec/tlv/LwM2mNodeTlvDecoder.java
@@ -33,10 +33,10 @@ import org.eclipse.leshan.core.node.LwM2mPath;
 import org.eclipse.leshan.core.node.LwM2mResource;
 import org.eclipse.leshan.core.node.LwM2mSingleResource;
 import org.eclipse.leshan.core.node.codec.CodecException;
-import org.eclipse.leshan.tlv.Tlv;
-import org.eclipse.leshan.tlv.Tlv.TlvType;
-import org.eclipse.leshan.tlv.TlvDecoder;
-import org.eclipse.leshan.tlv.TlvException;
+import org.eclipse.leshan.core.tlv.Tlv;
+import org.eclipse.leshan.core.tlv.TlvDecoder;
+import org.eclipse.leshan.core.tlv.TlvException;
+import org.eclipse.leshan.core.tlv.Tlv.TlvType;
 import org.eclipse.leshan.util.Hex;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
