+++ leshan-core/src/main/java/org/eclipse/leshan/core/tlv/Tlv.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.tlv;
+package org.eclipse.leshan.core.tlv;
 
 import java.util.Arrays;
