+++ leshan-core/src/main/java/org/eclipse/leshan/core/tlv/TlvEncoder.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.tlv;
+package org.eclipse.leshan.core.tlv;
 
 import java.nio.Buffer;
 import java.nio.ByteBuffer;
