+++ leshan-core/src/test/java/org/eclipse/leshan/tlv/TlvDecoderTest.java
@@ -21,6 +21,10 @@ import java.nio.ByteBuffer;
 import java.util.Arrays;
 
 import org.eclipse.leshan.core.node.ObjectLink;
+import org.eclipse.leshan.core.tlv.Tlv;
+import org.eclipse.leshan.core.tlv.TlvDecoder;
+import org.eclipse.leshan.core.tlv.TlvEncoder;
+import org.eclipse.leshan.core.tlv.TlvException;
 import org.eclipse.leshan.util.Hex;
 import org.junit.Test;
 import org.slf4j.Logger;
