+++ leshan-core/src/test/java/org/eclipse/leshan/tlv/TlvEncoderTest.java
@@ -21,7 +21,11 @@ import static org.junit.Assert.assertEquals;
 import java.nio.ByteBuffer;
 import java.util.Date;
 
-import org.eclipse.leshan.tlv.Tlv.TlvType;
+import org.eclipse.leshan.core.tlv.Tlv;
+import org.eclipse.leshan.core.tlv.TlvDecoder;
+import org.eclipse.leshan.core.tlv.TlvEncoder;
+import org.eclipse.leshan.core.tlv.TlvException;
+import org.eclipse.leshan.core.tlv.Tlv.TlvType;
 import org.junit.Test;
 
 public class TlvEncoderTest {
