+++ leshan-server-cf/src/test/java/org/eclipse/leshan/server/californium/request/CoapRequestBuilderTest.java
@@ -47,11 +47,11 @@ import org.eclipse.leshan.core.request.ReadRequest;
 import org.eclipse.leshan.core.request.WriteAttributesRequest;
 import org.eclipse.leshan.core.request.WriteRequest;
 import org.eclipse.leshan.core.request.WriteRequest.Mode;
+import org.eclipse.leshan.core.tlv.Tlv;
+import org.eclipse.leshan.core.tlv.TlvDecoder;
+import org.eclipse.leshan.core.tlv.Tlv.TlvType;
 import org.eclipse.leshan.server.registration.Registration;
 import org.eclipse.leshan.server.registration.Registration.Builder;
-import org.eclipse.leshan.tlv.Tlv;
-import org.eclipse.leshan.tlv.Tlv.TlvType;
-import org.eclipse.leshan.tlv.TlvDecoder;
 import org.junit.BeforeClass;
 import org.junit.Test;
